Removed boolean flags from adapter
