+++ app/src/main/java/com/example/next_show/adapters/ShowAdapter.java
@@ -12,8 +12,10 @@ import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.next_show.R;
-import com.example.next_show.data.NavigationInterface;
+import com.example.next_show.navigators.NavigationInterface;
 import com.example.next_show.models.Show;
+
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
@@ -21,19 +23,22 @@ public class ShowAdapter extends RecyclerView.Adapter<ShowAdapter.ViewHolder> {
     // instance variables
     private Context context;
     private List<Show> shows;
-    private boolean fromSavedFragment;
+    private NavigationInterface nav;
 
-    public ShowAdapter (Context context, List<Show> shows, Boolean fromSavedFragment){
+    // constant to track if adapter is being called from saved
+    public static final String SAVED_FRAGMENT = "SavedFragment";
+
+    public ShowAdapter (Context context, List<Show> shows, NavigationInterface nav){
         this.context = context;
         this.shows = shows;
-        this.fromSavedFragment = fromSavedFragment;
+        this.nav = nav;
     }
 
     @NonNull
     @Override
     public ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(context).inflate(R.layout.item_show, parent, false);
-        return new ViewHolder(view);
+        return new ViewHolder(view, nav);
     }
 
     @Override
@@ -54,36 +59,28 @@ public class ShowAdapter extends RecyclerView.Adapter<ShowAdapter.ViewHolder> {
 
     public void addAll(List<Show> list) {
         List<Show> modifiedShows = list;
-        if(fromSavedFragment){
+        if(nav.getSource().equals(SAVED_FRAGMENT)){
             modifiedShows = Show.fromParseShows(list);
         }
         shows.addAll(modifiedShows);
         notifyDataSetChanged();
     }
 
-    class NavigateFeedToDetail implements NavigationInterface {
-        public void navigate(View v, Bundle b){
-            Navigation.findNavController(v).navigate(R.id.action_feedFragment_to_showDetailFragment, b);
-        }
-    }
-
-    class NavigateSavedToDetail implements NavigationInterface {
-        public void navigate(View v, Bundle b){
-            Navigation.findNavController(v).navigate(R.id.action_savedFragment_to_showDetailFragment, b);
-        }
-    }
-
     class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
         // view elements
         private TextView tvShowBody;
         private TextView tvShowTitle;
+        private NavigationInterface navigator;
 
-        public ViewHolder(@NonNull View itemView) {
+        public ViewHolder(@NonNull View itemView, NavigationInterface navigator) {
             super(itemView);
             setUpView();
 
             // add this as click listener
             itemView.setOnClickListener(this);
+
+            // set navigator
+            this.navigator = navigator;
         }
 
         private void setUpView() {
@@ -101,15 +98,11 @@ public class ShowAdapter extends RecyclerView.Adapter<ShowAdapter.ViewHolder> {
 
                 // create new bundle of arguments to pass into Navigation
                 Bundle bundle = new Bundle();
-                bundle.putParcelable("Show", show);
+                Boolean fromSavedFragment = navigator.getSource().equals(SAVED_FRAGMENT);
                 bundle.putBoolean("savedBool", fromSavedFragment);
+                bundle.putParcelable("Show", show);
 
-                NavigationInterface navigator;
-                if (fromSavedFragment) {
-                    navigator = new NavigateSavedToDetail();
-                } else {
-                    navigator = new NavigateFeedToDetail();
-                }
+                // navigate to next fragment
                 navigator.navigate(v, bundle);
             }
         }
