--- app/src/main/java/com/example/next_show/data/NavigationInterface.java
@@ -1,8 +1,0 @@
-package com.example.next_show.data;
-
-import android.os.Bundle;
-import android.view.View;
-
-public interface NavigationInterface {
-    public void navigate(View v, Bundle b);
-}
