--- app/src/main/java/com/example/next_show/data/SavedFragmentCallback.java
@@ -1,2 +1,0 @@
-package com.example.next_show.data;public interface SavedFragmentCallback {
-}
