+++ app/src/main/java/com/example/next_show/navigators/NavigationInterface.java
@@ -1,0 +1,9 @@
+package com.example.next_show.navigators;
+
+import android.os.Bundle;
+import android.view.View;
+
+public interface NavigationInterface {
+    public void navigate(View v, Bundle b);
+    public String getSource();
+}
