+++ app/src/main/java/com/example/next_show/navigators/SavedFragmentCallback.java
@@ -1,0 +1,5 @@
+package com.example.next_show.navigators;
+
+public interface SavedFragmentCallback {
+
+}
